app/requirements.txt
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.1
bleach==3.1.5
blis==0.4.1
Bottleneck==1.3.2
catalogue==1.0.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
cymem==2.0.3
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
fastai==1.0.61
fastprogress==0.2.3
Flask==1.1.2
future==0.18.2
idna==2.9
importlib-metadata==1.6.1
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
joblib==0.15.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.3
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
kiwisolver==1.2.0
lxml==4.5.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numexpr==2.7.1
numpy==1.18.5
nvidia-ml-py3==7.352.0
opencv-python==4.2.0.34
packaging==20.4
pandas==1.0.4
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
Pillow==7.1.2
plac==1.1.3
preshed==3.0.2
prometheus-client==0.8.0
prompt-toolkit==3.0.5
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
requests==2.23.0
scikit-learn==0.23.1
scipy==1.4.1
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.0.1
spacy==2.2.4
srsly==1.0.2
terminado==0.8.3
testpath==0.4.4
thinc==7.4.0
threadpoolctl==2.1.0
torch==1.5.0+cpu
torchvision==0.6.0+cpu
tornado==6.0.4
tqdm==4.46.1
traitlets==4.3.3
urllib3==1.25.9
wasabi==0.6.0
wcwidth==0.2.3
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
zipp==3.1.0
